+++ ____old/applications/autoencoders.py
