+++ ____old/applications/gans/vanilla_gan.py
@@ -1,37 +1,37 @@
 from ..applications import Application
 from ..classification import Classifier
-from ...core.config import Config, Ref
-from ...core.templates import Layer, MultiInputLayer
-from ...components import (
+from ...core.config import Config
+from ...core.templates import Layer
+from ....deeplay.components import (
     SpatialBroadcastDecoder2d,
     PositionalEncodingSinusoidal2d,
-    ImageToVectorEncoder,
-    Concatenate,
     PositionalEncodingLinear2d,
     ImageRegressionHead,
-    MultiLayerPerceptron,
-    CategoricalClassificationHead,
 )
 import torch
 import torch.nn as nn
 from torch.optim import Adam
 
 
-class ClassConditionedGAN(Application):
+class VanillaGAN(Application):
     @staticmethod
     def defaults():
         return (
             Config()
-            .num_classes(10)
             .hidden_dim(10)
-            .embedding(
-                nn.Embedding,
-                num_embeddings=Ref("num_classes"),
-                embedding_dim=Ref("num_classes", lambda x: x // 2),
+            .generator(Layer("backbone") >> Layer("head"))
+            .generator.backbone(SpatialBroadcastDecoder2d)
+            .generator.backbone.encoding(PositionalEncodingLinear2d)
+            .generator.head(ImageRegressionHead)
+            .discriminator(Classifier)
+            .discriminator.num_classes(1)
+            .discriminator.on_first_forward(
+                "generator.backbone.output_size", lambda _, x: x.shape[2:]
             )
             .discriminator_loss(nn.MSELoss)
+            .generator_loss(nn.MSELoss)
             .discriminator_optimizer(Adam, lr=1e-4, betas=(0.5, 0.999))
-            .generator_optimizer(Adam, lr=1e-4, betas=(0.5, 0.999))
+            .generator_optimizer(Adam, lr=2e-4, betas=(0.5, 0.999))
         )
 
     def __init__(
@@ -61,9 +61,9 @@ class ClassConditionedGAN(Application):
 
         self.generator = self.new("generator")
         self.discriminator = self.new("discriminator")
-        self.embedding = self.new("embedding")
 
         self.discriminator_loss = self.new("discriminator_loss")
+        self.generator_loss = self.new("generator_loss")
 
     def configure_optimizers(self):
         return (
@@ -77,45 +77,28 @@ class ClassConditionedGAN(Application):
             ),
         )
 
-    def sample_noise(self, size):
-        batch_size, *spatials = size
-        return torch.randn(batch_size, self.hidden_dim, *spatials).to(self.device)
+    def sample_noise(self, batch_size):
+        return torch.randn(batch_size, self.hidden_dim).to(self.device)
 
-    def sample_conditioned_latent(self, x, z=None):
-        if z is None:
-            z = self.sample_noise(x.size())
-            z = z.to(self.device).float()
-
-        embed = self.embedding(x.long())
-        conditioned_latent = torch.cat([embed, z], dim=1)
-        return conditioned_latent
-
-    def forward(self, x, z=None):
-        image = self.generate(x, z=z)
-        valid = self.discriminate(image, x)
-        return image, valid
-
-    def discriminate(self, x, condition):
-        embed = self.embedding(condition)
-        return self.discriminator(x, embed)
-
-    def generate(self, condition, z=None):
-        conditioned_latent = self.sample_conditioned_latent(condition, z=z)
-        return self.generator(conditioned_latent)
+    def forward(self, x=None):
+        if x is None:
+            x = self.sample_noise(1)
+            x = x.to(self.device).float()
+        return self.generator(x)
 
     def training_step(self, batch, batch_idx, optimizer_idx):
-        x, condition = batch
-
-        z = self.sample_noise(condition.size())
-        z = z.to(self.device).float()
+        x = self._get_image(batch)
+        z = self.sample_noise(x.size(0))
 
         if optimizer_idx == 0:
             # store generated images for the discriminator optimizer
-            y_hat = self.generate(condition, z=z)
+            self._generated_imgs = self(z)
 
             valid = torch.ones(x.size(0), 1).type_as(x)
 
-            g_loss = self.discriminator_loss(self.discriminate(y_hat, condition), valid)
+            g_loss = self.generator_loss(
+                self.discriminator(self._generated_imgs), valid
+            )
 
             self.log("g_loss", g_loss, prog_bar=True, on_epoch=True)
 
@@ -124,13 +107,16 @@ class ClassConditionedGAN(Application):
         if optimizer_idx == 1:
             valid = torch.ones(x.size(0), 1).type_as(x)
             fake = torch.zeros(x.size(0), 1).type_as(x)
-            fake_x = self.generate(condition, z=z).detach()
-            real_loss = self.discriminator_loss(self.discriminate(x, condition), valid)
+
+            real_loss = self.discriminator_loss(self.discriminator(x), valid)
             fake_loss = self.discriminator_loss(
-                self.discriminate(fake_x, condition), fake
+                self.discriminator(self.generator(z).detach()), fake
             )
             d_loss = (real_loss + fake_loss) / 2
             self.log("d_loss", d_loss, prog_bar=True, on_epoch=True)
+            # if d_loss.item() < 0.25:
+            #     return d_loss * 0
+
             return d_loss
 
     def _get_image(self, batch):
