+++ ____old/components/bottlenecks.py
