+++ ____old/components/encoderdecoders/encoderdecoder.py
