+++ ____old/components/encodings/positional/linear.py
