+++ ____old/components/encodings/positional/sinusoidal.py
