+++ ____old/components/encodings/positional/utils.py
