+++ ____old/components/skip.py
